added to my cart

+++ src/main/java/com/deskbudd/Repo/UserRepo.java
@@ -5,7 +5,7 @@ import com.deskbudd.models.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface UserRepo extends JpaRepository<UserModel, Integer> {
+public interface UserRepo extends JpaRepository<UserModel, Long> {
 
     public UserModel findByEmail(String email);
 }
